// apps/livescore/src/presenter/theme/types.ts
export type TextKeys =
  | '$label01'
  | '$label02'
  | '$label03'
  | '$label01-inverted'
  | '$label02-inverted'
  | '$label03-inverted';

type UIKeys =
  | '$ui01'
  | '$ui02'
  | '$ui03'
  | '$red'
  | '$orange'
  | '$yellow'
  | '$green'
  | '$teal'
  | '$blue'
  | '$indigo'
  | '$purple'
  | '$pink'
  | '$shadow'
  | '$ui01-inverted'
  | '$ui02-inverted'
  | '$ui03-inverted'
  | '$red-inverted'
  | '$orange-inverted'
  | '$yellow-inverted'
  | '$green-inverted'
  | '$teal-inverted'
  | '$blue-inverted'
  | '$indigo-inverted'
  | '$purple-inverted'
  | '$pink-inverted'
  | '$shadow-inverted';

type BrandKeys =
  | '$brand01'
  | '$brand02'
  | '$brand03'
  | '$brand01-inverted'
  | '$brand02-inverted'
  | '$brand03-inverted';

export type UIColors = Pick<typeof THEME_COLORS, UIKeys>;
export type TextColors = Pick<typeof THEME_COLORS, TextKeys>;
export type BrandColors = Pick<typeof THEME_COLORS, BrandKeys>;

export type Colors = typeof THEME_COLORS;

export type UIKitBaseProps = {
  height?: number | string;
  flex?: number;
  borderRadius?: number | string;
  marginRight?: number | string;
  marginLeft?: number | string;
  marginTop?: number | string;
  marginBottom?: number | string;
  paddingTop?: number | string;
  paddingBottom?: number | string;
  paddingLeft?: number | string;
  paddingRight?: number | string;
  paddingVertical?: number | string;
  paddingHorizontal?: number | string;
  marginVertical?: number | string;
  marginHorizontal?: number | string;
  margin?: number | string;
  padding?: number | string;
  opacity?: number;
  borderColor?: UIKeys | 'transparent';
  borderWidth?: number | string;
  backgroundColor?: UIKeys | BrandKeys | 'transparent';
  width?: number | string;
  top?: number | string;
  left?: number | string;
  right?: number | string;
  bottom?: number | string;
  zIndex?: number;
  justifyContent?:
    | 'flex-start'
    | 'flex-end'
    | 'center'
    | 'space-between'
    | 'space-around'
    | 'space-evenly';
  alignItems?: 'flex-start' | 'flex-end' | 'center' | 'stretch' | 'baseline';
  flexDirection?: 'row' | 'column' | 'row-reverse' | 'column-reverse';
  flexWrap?: 'wrap' | 'nowrap' | 'wrap-reverse';
  alignContent?:
    | 'flex-start'
    | 'flex-end'
    | 'center'
    | 'stretch'
    | 'space-between'
    | 'space-around';
  alignSelf?:
    | 'auto'
    | 'flex-start'
    | 'flex-end'
    | 'center'
    | 'stretch'
    | 'baseline';
  position?: 'absolute' | 'relative';
  overflow?: 'visible' | 'hidden' | 'scroll';
  color?: TextKeys | BrandKeys | 'transparent';
  style?: AnyObject;
  textAlign?: 'left' | 'right' | 'center' | 'justify';
  textDecoration?: 'none' | 'underline' | 'line-through';
  textTransform?: 'none' | 'capitalize' | 'uppercase' | 'lowercase';
};

export type UIKitBaseViewProps = UIKitBaseProps;

export type UIKitBaseTextProps = TextProps &
  Omit<UIKitBaseProps, 'textAlign'> & {
    textAlign?: 'left' | 'right' | 'center' | 'justify';
    color?: TextKeys | BrandKeys;
    font?: Typography;
    textDecoration?: 'none' | 'underline' | 'line-through';
  };

export type TextInputProps = RNTextInputProps & UIKitBaseProps;

export type ButtonBaseProps = UIKitBaseProps &
  TouchableOpacityProps & {
    /**
     * Main colors
     */
    backgroundColor?: keyof BrandColors | keyof UIColors;
    buttonStyle?: ViewStyle;
    /**
     * Label
     */
    label: string;
    labelColor?: keyof TextColors;
    labelStyle?: TextStyle;

    disabled?: boolean;
    loading?: boolean;
    visible?: boolean;
  };